requirements.txt
absl-py==0.12.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
dataclass-csv==1.3.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.1
idna==2.10
joblib==1.0.1
Markdown==3.3.4
numpy==1.20.2
oauthlib==3.1.0
Pillow==8.2.0
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.23.2
scipy==1.6.3
six==1.16.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
torch==1.7.1
torchsummary==1.5.1
torchvision==0.8.2
tqdm==4.49.0
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==1.0.1
python-dotenv==0.17.1
memory_profiler